[REFACTOR] 캡슐화 리팩토링
